requirements.txt
torch
librosa
transformers
datasets
scikit-learn
numpy
pyannote.metrics
